// backend/src/types/traefik.ts
export interface TraefikRouter {
  entryPoints: string[];
  rule: string;
  service: string;
  tls?: {
    certResolver?: string;
  } | boolean;
  middlewares?: string[];
}

export interface TraefikService {
  loadBalancer: {
    servers: Array<{
      url: string;
    }>;
    healthCheck?: {
      interval?: string;
      path?: string;
      timeout?: string;
    };
    serversTransport?: string;
  };
}

export interface TraefikMiddleware {
  errors?: {
    query: string;
    service: string;
    status: string[];
  };
  rateLimit?: {
    average: number;
    burst: number;
  };
  headers?: {
    customRequestHeaders?: Record<string, string>;
    customResponseHeaders?: Record<string, string>;
  };
  redirectRegex?: {
    permanent: boolean;
    regex: string;
    replacement: string;
  };
}

export interface TraefikConfig {
  http: {
    routers: Record<string, TraefikRouter>;
    services: Record<string, TraefikService>;
    middlewares: Record<string, TraefikMiddleware>;
    serversTransports?: Record<string, any>;
  };
}